Allow running a command other than `rspec`
Sometimes tests have complicated setup, such as initializing a
database or moving some files around. We now allow running a script
that does all of the hard work and calls RSpec for us.

One ugly thing is that we split the argument on whitespace, so scripts
with meaningful whitespace in the name won't work. Hopefully this
won't be a problem, as the user can call the script anything they
want.

Closes #9